chore: Achieve clippy strict pedantic cleanliness
- Remove unused imports: FieldSource, StreamExt
- Remove unused constants: STATE_COMPLETE, STATE_ERROR
- Remove unused struct fields and dead code methods
- Fix variable mutability issues (remove unnecessary mut keywords)
- Fix logic quality: redundant closures, unnecessary casts, nested ifs
- Add 22 missing documentation comments to operator struct fields
- Fix test structs with #[allow(dead_code)] for serde-only fields
- Add #[allow(clippy::large_enum_variant)] for unavoidable TLS crypto types
- Add #[allow(clippy::too_many_arguments)] for 9-param streaming_query
- Replace useless format! calls with .to_string()
- Fix empty format strings and io::Error patterns

Result: 0 clippy errors with -D warnings flag
All 158 library tests passing

+++ src/protocol/decode.rs
@@ -156,7 +156,7 @@ fn decode_data_row(data: &[u8]) -> io::Result<BackendMessage> {
         if offset + 4 > data.len() {
             return Err(io::Error::new(io::ErrorKind::UnexpectedEof, "field length"));
         }
-        let field_len = i32::from_be_bytes([data[offset], data[offset+1], data[offset+2], data[offset+3]]) as i32;
+        let field_len = i32::from_be_bytes([data[offset], data[offset+1], data[offset+2], data[offset+3]]);
         offset += 4;
 
         let field = if field_len == -1 {
